move AVSAudioLevel to separate file
